requirements.txt
Flask
Pillow
Flask-Login
Flask-OIDC
python-dotenv
requests
oauthlib
